Don't follow web links by default (check, serve) (#22)
When running `nix run .#check` or `nix run .#serve`, don't linkcheck external links.

The reason is that several of HashiCorp's resources that we link to serve 429s
that are supposed to mean Too Many Requests, but they serve them 100% of the
time to our linkchecker.

Instead, run `nix run .#deepcheck` to follow web links.

See #21.